+++ paddle/phi/kernels/funcs/cross_entropy.cu
@@ -14,9 +14,9 @@ limitations under the License. */
 
 #include "paddle/phi/kernels/funcs/cross_entropy.h"
 
-#include "paddle/fluid/platform/device/gpu/gpu_dnn.h"
 #include "paddle/phi/backends/gpu/gpu_context.h"
 #include "paddle/phi/backends/gpu/gpu_device_function.h"
+#include "paddle/phi/backends/gpu/gpu_dnn.h"
 #include "paddle/phi/backends/gpu/gpu_primitives.h"
 #include "paddle/phi/core/utils/data_type.h"
 #include "paddle/phi/kernels/funcs/math.h"
