+++ paddle/phi/kernels/gpu/instance_norm_utils.h
@@ -26,7 +26,7 @@
 namespace cub = hipcub;
 #endif
 
-#include "paddle/fluid/platform/device/gpu/gpu_dnn.h"
+#include "paddle/phi/backends/gpu/gpu_dnn.h"
 
 namespace phi {
